Implement type filters
Type filters are applied to every type before a variable is shown in the
variable view. They can be used to make the variables more readable.

Each filter consists of a regex pattern and what it should be replaced
by.

Example settings:
"type_filters": [
  {
    "pattern": " >",
    "replace": ">"
  },
  {
    "pattern": ", *std::allocator(<(?>[^<>]|(?1))*>)",
    "replace": ""
  }
]

If the regex module is available it is used instead of the re module.
This enables features like subroutine calls to search for matching
brackets as in the second example pattern.