Update Readme.creole
I'm somewhat of a Sublime novice, and this confused me. If you just type GDB into the command palette, the first option that comes up is Set Syntax: GDB. After playing it with a bit, I think it is intended to open the SublimeGDB Default Preferences, and this verbiage makes it more clear.